Load multiple markets simultaneously during startup (#16)
* load multiple markets at once. For large amounts of markets lots and lots of getAccountInfo's are being called simultaneously. Condense this into smaller GMA calls.

The associated account tokens are passed to the dex instructions but as it happens they don't actually need to be created for the market to be cranked. We can just get the address without calling the rpc to check if they exist or create them.

before: each market had to call getAccountInfo to get the account Info then 2 more calls to getAccountInfo to get the decimals for the baseMint and QuoteMint - even if we had already fetched these baseMints and quoteMints we would keep re-fetching them for each market that had them. There was also a further call to getAccountInfo to check if the associated token account existed for each baseMint token which would then create the account if it does not exist.

At least in my test environment, loading many markets was causing me RPC errors due to the spraying of all these getAccountInfo calls at the same time. This would happen a relatively small number of markets (around 25) but now with these changes I am able to load up to 325 markets at once (still testing limits)

---------

Co-authored-by: solpkr <solpkr1@gmail.com>